update replication component_update, entity_despawn, component_remove
