+++ benches/message.rs
@@ -11,9 +11,10 @@ use divan::{AllocProfiler, Bencher};
 use lightyear::client::sync::SyncConfig;
 use lightyear::prelude::client::{InterpolationConfig, PredictionConfig};
 use lightyear::prelude::{client, server, MessageRegistry, Tick, TickManager};
-use lightyear::prelude::{ClientId, NetworkTarget, SharedConfig, TickConfig};
+use lightyear::prelude::{ClientId, SharedConfig, TickConfig};
 use lightyear::server::input::InputBuffers;
 use lightyear::shared::replication::components::Replicate;
+use lightyear::shared::replication::network_target::NetworkTarget;
 use lightyear_benches::local_stepper::{LocalBevyStepper, Step as LocalStep};
 use lightyear_benches::protocol::*;
