+++ lightyear/src/client/message.rs
@@ -12,7 +12,7 @@ use crate::client::connection::ConnectionManager;
 use crate::client::events::MessageEvent;
 use crate::client::networking::is_connected;
 use crate::packet::message::SingleData;
-use crate::prelude::{ChannelDirection, ChannelKind, Message, NetworkTarget};
+use crate::prelude::{ChannelDirection, ChannelKind, Message};
 use crate::protocol::message::{MessageKind, MessageRegistry};
 use crate::protocol::registry::NetId;
 use crate::protocol::BitSerializable;
@@ -20,6 +20,7 @@ use crate::serialize::reader::ReadBuffer;
 use crate::serialize::writer::WriteBuffer;
 use crate::serialize::RawData;
 use crate::shared::ping::message::{Ping, Pong, SyncMessage};
+use crate::shared::replication::network_target::NetworkTarget;
 use crate::shared::replication::{ReplicationMessage, ReplicationMessageData};
 use crate::shared::sets::{ClientMarker, InternalMainSet};
